+++ lib/PuppeteerSharp.DevicesFetcher/Models.cs
@@ -1,54 +1,7 @@
-﻿using System.Collections.Generic;
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 
 namespace PuppeteerSharp.DevicesFetcher
 {
-    public class DevicePayload
-    {
-        public string Type { get; set; }
-
-        public string UserAgent { get; set; }
-
-        public ViewportPayload Vertical { get; set; }
-
-        public ViewportPayload Horizontal { get; set; }
-
-        public double DeviceScaleFactor { get; set; }
-
-        public HashSet<string> Capabilities { get; set; }
-    }
-
-    public class ViewportPayload
-    {
-        [JsonProperty("width")]
-        public double Width { get; set; }
-
-        [JsonProperty("height")]
-        public double Height { get; set; }
-    }
-
-    public class OutputDevice
-    {
-        public string Name { get; set; }
-        public string UserAgent { get; set; }
-        public OutputDeviceViewport Viewport { get; set; }
-
-        public class OutputDeviceViewport
-        {
-            public double Width { get; set; }
-
-            public double Height { get; set; }
-
-            public double DeviceScaleFactor { get; set; }
-
-            public bool IsMobile { get; set; }
-
-            public bool HasTouch { get; set; }
-
-            public bool IsLandscape { get; set; }
-        }
-    }
-
     public class RootObject
     {
         [JsonProperty("extensions")]
@@ -75,8 +28,7 @@ namespace PuppeteerSharp.DevicesFetcher
             [JsonProperty("user-agent")]
             public string UserAgent { get; set; }
             [JsonProperty("type")]
-            public string Type { get; set; }
-            
+            public string Type { get; set; }            
         }
 
         public class Screen
